Add a healthcheck to the webserver
This change is to reduce downtime when the issue described in https://issues.apache.org/jira/browse/AIRFLOW-1235 is encountered. (The master gunicorn process dies but the airflow CLI does not fail or restart, resulting in a dead webserver). Currently when this condition is met, the webserver has to be manually killed or restarted.

This change will poll for the airflow-webserver pid file and the webserver container will restart if it is not present (which is the case when the gunicorn master dies)

Related to #125