Add support for restart on watched files.
If the env variable `WATCH_FILES` is set, use the
[entr(1)](http://eradman.com/entrproject/) tool to wrap the entrypoint
script to automatically restart the airflow process if a change is
detected in any of the watched files.

The use case here is: a sidecar container syncs the dags directory from
an external source, and that sync updates
/usr/local/airflow/dags/requirements.txt: in this case we would want to
restart the airflow process after re-processing the requirements.